==> R/01_construct-and-calibrate-model.Rmd <==
---
title: "Calibrate a Model of U.S. Health Insurance Markets"
output:
  pdf_document: default
  html_notebook: default
editor_options:
  chunk_output_type: console
---
The objective of this document is to construct and calibrate a model of U.S. insurance markets.

# Prepare Data

```{r, echo = FALSE, warning = FALSE}
library(here)
source(here::here("/R/manifest.R"))
library(Matrix)
```

```{r}
create_sipp_data = FALSE
if (create_sipp_data) source(here("R/read-and-tidy-SIPP-data.R"))

df_sipp_full <- read_rds(here("input/sipp/01_sipp-tidy_v1-0.rds"))

# Get survey weight (use value from first month)
df_w <- 
  df_sipp_full %>% 
  mutate(idnumber = id) %>% 
  group_by(idnumber) %>% 
  mutate(year = ifelse(swave ==1 , 2014, 2015)) %>% 
  mutate(month = ifelse(swave ==1 , monthcode, monthcode+12)) %>% 
  filter(month==1) %>% 
  select(idnumber,weight = wpfinwgt) %>% 
  ungroup() 

df_sipp <- 
  df_sipp_full %>%    
  filter(age < 65 & age > 18) %>% 
  mutate(year = ifelse(swave ==1 , 2014, 2015)) %>% 
  mutate(month = ifelse(swave ==1 , monthcode, monthcode+12)) %>% 
  mutate(idnumber = id) %>% 
  mutate(insurance_type = factor(hicov, labels = insurance_sipp_lut)) %>% 
  select(idnumber,month,insurance_type,expansion_state, sex, race, state)  %>% 
  # !!!! Note expansion state is time-varying if someone moves. 
  group_by(idnumber) %>% 
  mutate(expansion_state = max(expansion_state)) %>% 
  ungroup()  %>% 
  left_join(df_w,"idnumber")

ex_ante <- 
  df_sipp %>% 
  filter(month==1) %>% 
  group_by(insurance_type) %>% 
  summarise(n = sum(weight,na.rm=TRUE)) %>% 
  ungroup() %>% 
  mutate(pct = n/sum(n)) 

ex_ante %>%  
  write_rds(here("output/ex-ante-overall-population/ex-ante-distribution.rds"))

ex_post <- 
  df_sipp %>% 
  filter(month==13) %>% 
  group_by(insurance_type) %>% 
  summarise(n = sum(weight,na.rm=TRUE)) %>% 
  ungroup() %>% 
  mutate(pct = n/sum(n)) 

ex_post %>%  
  write_rds(here("output/ex-ante-overall-population/ex-post-distribution.rds"))

# Targets are based on ACS tables for 19-64 year olds in ACS data as constructed at http://statehealthcompare.shadac.org/                    
# Respondents are asked about their health insurance coverage at the time of the interview and may select multiple types of coverage. SHADAC's primary source of coverage hierarchy is applied to assign one coverage type to each respondent. If multiple sources of coverage were reported for an observation, primary source of coverage was assigned in the following order: 1) Medicare (for people age 19 or older); 2) employer-sponsored insurance (ESI), TRICARE or other military health care, or VA; 3) Medicaid; 4) Individual coverage; and 5) Medicare (for people age 18 or under). Private coverage includes employer (plus TRICARE and VA) and individual coverage. Public coverage includes Medicaid (plus CHIP and state-specific public programs) and Medicare. The margin of error (MOE) represents the uncertainty of an estimate due to sampling variability; the calculated MOE is 95%. N/A indicates that data were not available or suppressed because the number of sample cases was too small.


lst_targets <- list(ESI = c(0.64,	0.635,	0.629,	0.624),
                    NG = c(0.082,	0.086,	0.092,	0.092),
                    PUB = c(0.155,	0.158,	0.158,	0.154),
                    UNIN = c(0.124,	0.122,	0.12,	0.131))

# Set Up the Multi-State Data
ls_ms <- 
  df_sipp %>% 
  mutate(insurance_type = paste0(insurance_type)) %>% 
  mutate(constant = 1) %>% 
  prepare_multistate_data(idvar = idnumber,
                          timevar = month, 
                          statevar = insurance_type) 

categories <- names(ls_ms$trans_mat)

fit_transition <- map(categories,
                     ~ (
                       fit_multistate_model(
                         df = ls_ms$df_ms %>% pluck(.x),
                         tmat = ls_ms$trans_mat %>% pluck(.x),
                         fit_type = "km",
                         ff = Surv(Tstart, Tstop, status) ~ 1,
                         idvar = idnumber,
                         prediction_vals = data.frame(constant = 1)
                       ) 
                     )) %>% 
  set_names(categories)

trans_probs <- 
  categories %>% 
    map(~(
    get_cumHaz(dist = fit_transition[[1]]$fit_type ,
               ls_fit = fit_transition %>% pluck(.x), 
               tt = 1:24,
               lut = fit_transition %>% 
                 pluck(.x) %>% pluck("lut"))  %>% 
      mutate(trans = paste0("pstate",trans)) %>% 
      spread(trans,Haz) %>% 
      magrittr::set_names(c("time",fit_transition %>% pluck(.x) %>% 
                              pluck("lut") %>% pull(transition_type))) %>% 
      mutate(baseline = .x) %>% 
      janitor::clean_names()
  )) %>% 
  bind_rows() %>% 
  select(time,baseline,x01_esi_own,x02_priv_oth,x03_public,x04_uninsured) %>% 
  group_by(time) %>% 
  nest() %>% 
  mutate(rates = map(data,~(
    .x %>% as.data.frame() %>% 
      column_to_rownames(var = "baseline") %>% 
      mutate(x01_esi_own = ifelse(is.na(x01_esi_own),-(x02_priv_oth+x03_public+x04_uninsured),x01_esi_own)) %>% 
      mutate(x02_priv_oth = ifelse(is.na(x02_priv_oth),-(x01_esi_own+x03_public+x04_uninsured),x02_priv_oth)) %>% 
      mutate(x03_public = ifelse(is.na(x03_public),-(x01_esi_own+x02_priv_oth+x04_uninsured),x03_public)) %>% 
      mutate(x04_uninsured = ifelse(is.na(x04_uninsured),-(x01_esi_own+x02_priv_oth+x03_public),x04_uninsured)) %>% 
      set_names(gsub("^x","",names(.x)[-1])) %>% 
      set_rownames(gsub("^x","",names(.x)[-1])) %>% 
      as.matrix() 
  ))) %>% 
  mutate(prob = map2(rates,data,~(
    as.matrix(expm(.x)) %>% 
      data.frame() %>% 
      rownames_to_column(var = "baseline") %>% 
      set_names(gsub("^x","",names(.y))) %>% 
      tbl_df()
  )))
  
trans_probs %>% 
  write_rds(here("output/transition-probabilities/transition-probabilities-kaplan-meier.rds"))
```


```{r}
p <- lst_targets %>% transpose() %>% pluck(1) %>% unlist() #ex_ante$pct %>% as.matrix() %>% t()
R <- trans_probs %>% 
  filter(time == 13) %>% 
  pull(data) %>% pluck(1) %>% 
  select(-baseline) %>% 
  as.matrix()

p_ESI_NG <- R[1,2]
p_ESI_PUB <- R[1,3]
p_ESI_UNIN <- R[1,4]

p_NG_ESI <- R[2,1]
p_NG_PUB <- R[2,3]
p_NG_UNIN <- R[2,4]

p_PUB_ESI <- R[3,1]
p_PUB_NG <- R[3,2]
p_PUB_UNIN <- R[3,4]

p_UNIN_ESI <- R[4,1]
p_UNIN_NG <- R[4,2]
p_UNIN_PUB <- R[4,3]

ev <- c(p_ESI_NG = p_ESI_NG, p_ESI_PUB = p_ESI_PUB, p_ESI_UNIN = p_ESI_UNIN,
        p_NG_ESI = p_NG_ESI, p_NG_PUB = p_NG_PUB, p_NG_UNIN = p_NG_UNIN,
        p_PUB_ESI = p_PUB_ESI, p_PUB_NG = p_PUB_NG, p_PUB_UNIN = p_PUB_UNIN,
        p_UNIN_ESI = p_UNIN_ESI, p_UNIN_NG = p_UNIN_NG, p_UNIN_PUB = p_UNIN_PUB)

run_insurance_markov <- function(v_params) {
  with(as.list(v_params), {
    n_t <- 4
    v_n  <- c("ESI", "NG", "PUB", "UNIN") # the 4 states of the model: ESI, non-group, public, uninsured
    n_s <- length(v_n)              # number of insurance categories
  
    ####### INITIALIZATION ##########################################
    # create the cohort trace
    m_M <- matrix(NA, nrow = n_t + 1 , 
                  ncol = n_s,
                  dimnames = list(0:n_t, v_n))     # create Markov trace (n_t + 1 because R doesn't understand  Cycle 0)
    
    m_M[1, ] <- p                      # initialize Markov trace
    
    # create transition probability matrix for NO treatment
    m_P <- matrix(0,
                  nrow = n_s, 
                  ncol = n_s,
                  dimnames = list(v_n, v_n))
    # fill in the transition probability array
    ### From ESI
    m_P["ESI", "ESI"]  <- 1 - (p_ESI_NG + p_ESI_PUB + p_ESI_UNIN)
    m_P["ESI", "NG"] <- p_ESI_NG
    m_P["ESI", "PUB"] <- p_ESI_PUB
    m_P["ESI", "UNIN"] <- p_ESI_UNIN

    ### From Non-Group
    m_P["NG", "NG"]  <- 1 - (p_NG_ESI + p_NG_PUB + p_NG_UNIN)
    m_P["NG", "ESI"] <- p_NG_ESI
    m_P["NG", "PUB"] <- p_NG_PUB
    m_P["NG", "UNIN"] <- p_NG_UNIN
    
    ### From Non-Group
    m_P["PUB", "NG"]  <- p_PUB_NG
    m_P["PUB", "ESI"] <- p_PUB_ESI
    m_P["PUB", "PUB"] <- 1 - (p_PUB_ESI + p_PUB_NG + p_PUB_UNIN)
    m_P["PUB", "UNIN"] <- p_PUB_UNIN
    
    
    ### From UNINSURED
    m_P["UNIN", "NG"]  <- p_UNIN_NG
    m_P["UNIN", "ESI"] <- p_UNIN_ESI
    m_P["UNIN", "PUB"] <- p_UNIN_PUB
    m_P["UNIN", "UNIN"] <- 1 - (p_UNIN_ESI + p_UNIN_NG + p_UNIN_PUB)
    
    
    # check rows add up to 1
    if (!isTRUE(all.equal(as.numeric(rowSums(m_P)), as.numeric(rep(1, n_s))))) {
      stop("This is not a valid transition Matrix")
    }
    
        ############# PROCESS ###########################################
    
    for (t in 1:n_t){                              # throughout the number of cycles
      m_M[t + 1, ] <- m_M[t, ] %*% m_P           # estimate the Markov trace for cycle the next cycle (t + 1)
    }
    
    ####### EPIDEMIOLOGICAL OUTPUT  ###########################################
    #### Overall Survival (OS) ####
    #v_os <- 1 - m_M[, "UNIN"]                # calculate the overall survival (OS) probability for no treatment
    v_ESI <-  m_M[,"ESI"]
    v_NG <- m_M[,"NG"]
    v_PUB <- m_M[,"PUB"]
    v_UNIN <- m_M[,c("UNIN")]
    
    
    ####### RETURN OUTPUT  ###########################################
    #out <- list(coverage = m_M[2,])
    out <- list(ESI = m_M[-1,1],
                NG = m_M[-1,2],
                PUB = m_M[-1,3],
                UNIN = m_M[-1,4])
    
    return(out)
  }
  )
}

# Check that it works
v_params_test <- c(p_ESI_NG = 0.2, p_ESI_UNIN = 0.2)
run_insurance_markov(ev) # It works!
```

```{r}

# calibration functionality
library(lhs)
library(IMIS)
library(matrixStats) # package used for sumamry statistics

# visualization
library(plotrix)
library(psych)
```

```{r}
####################################################################
######  Specify calibration parameters  ######
####################################################################
# Specify seed (for reproducible sequence of random numbers)
set.seed(23)

prior_type = "uniform"

# number of random samples
n_resamp <- 5000

# names and number of input parameters to be calibrated
v_param_names <- c("p_ESI_NG","p_ESI_PUB","p_ESI_UNIN","p_NG_ESI","p_NG_PUB","p_NG_UNIN",
                   "p_PUB_ESI","p_PUB_NG","p_PUB_UNIN","p_UNIN_ESI", "p_UNIN_NG","p_UNIN_PUB")
n_param <- length(v_param_names)

# range on input search space
ev <- c(p_ESI_NG = R[1,2], p_ESI_PUB = R[1,3], p_ESI_UNIN =R[1,4], 
        p_NG_ESI = R[2,1], p_NG_PUB = R[2,3], p_NG_UNIN = R[2,4],
        p_PUB_ESI = R[3,1], p_PUB_NG = R[3,2], p_PUB_UNIN = R[3,4], 
        p_UNIN_ESI = R[4,1], p_UNIN_NG =R[4,2], p_UNIN_PUB = R[4,3]) # lower bound
lb <- c(p_ESI_NG = 0.001, p_ESI_PUB = 0.001, p_ESI_UNIN = 0.03, 
        p_NG_ESI = 0.05, p_NG_PUB = 0.001, p_NG_UNIN = 0.01,
        p_PUB_ESI = 0.01, p_PUB_NG = 0.01, p_PUB_UNIN = 0.02, 
        p_UNIN_ESI = 0.05, p_UNIN_NG = 0.01, p_UNIN_PUB = 0.05) # lower bound
ub <- c(p_ESI_NG = 0.15, p_ESI_PUB = 0.15, p_ESI_UNIN = 0.15, 
        p_NG_ESI = 0.30, p_NG_PUB = 0.20, p_NG_UNIN = 0.10,
        p_PUB_ESI = 0.15, p_PUB_NG = 0.1, p_PUB_UNIN = 0.3, 
        p_UNIN_ESI = 0.3, p_UNIN_NG = .3, p_UNIN_PUB = 0.3) # upper bound

# number of calibration targets
v_target_names <- c("ESI","NG","PUB","UNIN")
n_target     <- length(v_target_names)

### Calibration functions

#  Write function to sample from prior
sample_prior <- function(n_samp){
  m_lhs_unit   <- randomLHS(n = n_samp, k = n_param)
  m_param_samp <- matrix(nrow = n_samp, ncol = n_param)
  colnames(m_param_samp) <- v_param_names
  for (i in 1:n_param){
    
  if (prior_type=="beta") {
   m_param_samp[,i] <- qbeta(m_lhs_unit[,i],shape1 = ev[i]*100,shape2=100-ev[i]*100)
  } else {
    m_param_samp[, i] <- qunif(m_lhs_unit[,i],
                               min = lb[i],
                               max = ub[i])
    }
  }
  return(m_param_samp)
}


###  PRIOR  ### 
# Write functions to evaluate log-prior and prior

# function that calculates the log-prior
calc_log_prior <- function(v_params){
  if(is.null(dim(v_params))) { # If vector, change to matrix
    v_params <- t(v_params) 
  }
  n_samp <- nrow(v_params)
  colnames(v_params) <- v_param_names
  lprior <- rep(0, n_samp)
  for (i in 1:n_param){
    if (prior_type=="beta") {
      lprior <- lprior + dbeta(v_params[,i], shape1 = ev[i]*100, shape2 = 100-ev[i]*100,log = T)
    } else {
    lprior <- lprior + dunif(v_params[, i],
                             min = lb[i],
                             max = ub[i],
                             log = T)
    }

  }
  return(lprior)
}
calc_log_prior(v_params = ev)
calc_log_prior(v_params = sample_prior(10))

# function that calculates the (non-log) prior
calc_prior <- function(v_params) { 
  exp(calc_log_prior(v_params)) 
}
calc_prior(v_params = ev)
calc_prior(v_params = sample_prior(10))

###  LIKELIHOOD  ###
# Write functions to evaluate log-likelihood and likelihood

# function to calculate the log-likelihood
calc_log_lik <- function(v_params){
  # par_vector: a vector (or matrix) of model parameters 
  if(is.null(dim(v_params))) { # If vector, change to matrix
    v_params <- t(v_params) 
  }
  n_samp <- nrow(v_params)
  v_llik <- matrix(0, nrow = n_samp, ncol = n_target) 
  llik_overall <- numeric(n_samp)
  for(j in 1:n_samp) { # j=1
    jj <- tryCatch( { 
      ###   Run model for parametr set "v_params" ###
      model_res <- run_insurance_markov(v_params[j, ])
      
      ###  Calculate log-likelihood of model outputs to targets  ###
      # TARGET 1: Survival ("Surv")
      # log likelihood  
      v_llik[j, 1] <- sum(dnorm(x = lst_targets$ESI,
                                mean = model_res$ESI,
                                sd = 1*sqrt(lst_targets$ESI * (1-lst_targets$ESI))/1e2,
                                log = T))
      
      v_llik[j, 2] <- sum(dnorm(x = lst_targets$NG,
                                mean = model_res$NG,
                                sd = 1*sqrt(lst_targets$NG * (1-lst_targets$NG))/1e2,
                                log = T))
      v_llik[j, 3] <- sum(dnorm(x = lst_targets$PUB,
                                mean = model_res$PUB,
                                sd = 1*sqrt(lst_targets$PUB * (1-lst_targets$PUB))/1e2,
                                log = T))
      v_llik[j, 4] <- sum(dnorm(x = lst_targets$UNIN,
                                mean = model_res$UNIN,
                                sd = 1*sqrt(lst_targets$UNIN * (1-lst_targets$UNIN))/1e2,
                                log = T))      
      
      # OVERALL 
      llik_overall[j] <- sum(v_llik[j, ])
    }, error = function(e) NA) 
    if(is.na(jj)) { llik_overall <- -Inf }
  } # End loop over sampled parameter sets
  # return LLIK
  return(llik_overall)
}
calc_log_lik(v_params = ev)
calc_log_lik(v_params = sample_prior(10))


# function to calculate the (non-log) likelihood
calc_likelihood <- function(v_params){ 
  exp(calc_log_lik(v_params)) 
}
calc_likelihood(v_params = ev)
calc_likelihood(v_params = sample_prior(1000))

###  POSTERIOR  ###
# Write functions to evaluate log-posterior and posterior

# function that calculates the log-posterior
calc_log_post <- function(v_params) { 
  lpost <- calc_log_prior(v_params) + calc_log_lik(v_params)
  return(lpost) 
}
calc_log_post(v_params = ev)
calc_log_post(v_params = sample_prior(10))


# function that calculates the (non-log) posterior
calc_post <- function(v_params) { 
  exp(calc_log_post(v_params)) 
}
calc_post(v_params = ev)
calc_post(v_params = sample_prior(10))

####################################################################
######  Calibrate!  ######
####################################################################
# record start time of calibration
t_init <- Sys.time()

###  Bayesian calibration using IMIS  ###
# define three functions needed by IMIS: prior(x), likelihood(x), sample.prior(n)

# based on ACS tables for 19-64 year olds in ACS data as constructed at http://statehealthcompare.shadac.org/                    


prior <- calc_prior
likelihood <- calc_likelihood
sample.prior <- sample_prior

# run IMIS
fit_imis <- IMIS(B = 5000, # the incremental sample size at each iteration of IMIS
                 B.re = n_resamp, # the desired posterior sample size
                 number_k = 10, # the maximum number of iterations in IMIS
                 D = 0) 

# obtain draws from posterior
m_calib_res <- fit_imis$resample



# Calculate log-likelihood (overall fit) and posterior probability of each sample
m_calib_res <- cbind(m_calib_res, 
                      "Overall_fit" = calc_log_lik(m_calib_res[,v_param_names]),
                      "Posterior_prob" = calc_post(m_calib_res[,v_param_names]))

# normalize posterior probability
m_calib_res[,"Posterior_prob"] <- m_calib_res[,"Posterior_prob"]/sum(m_calib_res[,"Posterior_prob"])

# Calculate computation time
comp_time <- Sys.time() - t_init

####################################################################
######  Exploring posterior distribution  ######
####################################################################

m_calib_res %>% 
  tbl_df() %>%
  gather(parameter,value) %>% 
  filter(grepl("^p_",parameter)) %>% 
  ggplot(aes(x = value)) + geom_density() + 
  facet_wrap(~parameter, scales = "free",nrow = 7) + 
  theme_light() + 
  labs(x = "Density", y = "Parameter Value") + 
  ggtitle("Posterior Distribution of Transition Probabilities") + 
  ggsave("figures/01-posterior-distribution.png",width = 5, height = 8)

fit_imis$center %>% 
  tbl_df() %>% 
  gather(parameter,value) 

# Plot the 1000 draws from the posterior
v_post_color <- scales::rescale(m_calib_res[,"Posterior_prob"])
plot(m_calib_res[,3:4],
     xlim = c(lb[3], ub[3]), ylim = c(lb[4], ub[4]),
     xlab = v_param_names[3], ylab = v_param_names[4],
     col = scales::alpha("black", v_post_color))
# add center of Gaussian components
points(fit_imis$center[,3:4], col = "red", pch = 8)
legend("topright", c("Draws from posterior", "Center of Gaussian components"),
       col = c("black", "red"), pch = c(1, 8))

# Plot the 1000 draws from the posterior with marginal histograms
#pairs.panels(m_calib_res[,v_param_names])

# Compute posterior mean
v_calib_post_mean <- colMeans(m_calib_res[,v_param_names])
v_calib_post_mean

# Compute posterior median and 95% credible interval
m_calib_res_95cr <- colQuantiles(m_calib_res[,v_param_names], probs = c(0.025, 0.5, 0.975))
m_calib_res_95cr

# Compute maximum-a-posteriori (MAP) parameter set
v_calib_map <- m_calib_res[which.max(m_calib_res[,"Posterior_prob"]),]

### Plot model-predicted output at mode vs targets ###
v_out_best <- run_insurance_markov(v_calib_map[v_param_names])
v_out_mean <- run_insurance_markov(v_calib_post_mean)
v_out_orig <- run_insurance_markov(ev)

p_calib_df <- 
  bind_rows(
    lst_targets %>% bind_cols() %>% 
      mutate(year = 2015:2018) %>% 
      mutate(label = "ACS Target"),
    v_out_best %>% bind_cols() %>% 
      mutate(year = 2015:2018) %>% 
      mutate(label = "Calibrated"),
    v_out_orig %>% bind_cols() %>%
      mutate(year = 2015:2018) %>%
      mutate(label = "Uncalibrated")
    ) %>% 
      gather(type,value,-year,-label) 

library(patchwork)

p_calib_esi <- 
  p_calib_df %>% 
  filter(type=="ESI") %>% 
  ggplot(aes(x = year, y = value, colour = label)) + 
  geom_point() + 
  # facet_wrap(~type, scales = "free") + 
  theme_minimal() + 
  geom_line() +
  scale_y_continuous(limits = c(0.5,0.75)) + 
  scale_x_continuous(limits = c(2015,2019)) +
  labs(y = "Fraction of Population Covered", x = "Year") + 
  geom_dl(aes(label = label), method = list("last.polygons",cex =0.9,fontfamily = "Gill Sans")) + 
  theme(legend.position = "none") +
  scale_colour_brewer(palette="Dark2") +
  ggtitle("Employer-Sponsored Insurance")

p_calib_ng <- 
  p_calib_df %>% 
  filter(type=="NG") %>% 
  ggplot(aes(x = year, y = value, colour = label)) + 
  geom_point() + 
  # facet_wrap(~type, scales = "free") + 
  theme_minimal() + 
  geom_line() +
  scale_y_continuous(limits = c(0,0.25)) + 
  scale_x_continuous(limits = c(2015,2019)) +
  labs(y = "Fraction of Population Covered", x = "Year") + 
  geom_dl(aes(label = label), method = list("last.polygons",cex =0.9,fontfamily = "Gill Sans")) + 
  theme(legend.position = "none") +
  scale_colour_brewer(palette="Dark2") +
  ggtitle("Non-Group Private Insurance")


p_calib_pub <- 
  p_calib_df %>% 
  filter(type=="PUB") %>% 
  ggplot(aes(x = year, y = value, colour = label)) + 
  geom_point() + 
  # facet_wrap(~type, scales = "free") + 
  theme_minimal() + 
  geom_line() +
  scale_y_continuous(limits = c(0,0.25)) + 
  scale_x_continuous(limits = c(2015,2019)) +
  labs(y = "Fraction of Population Covered",x = "Year") + 
  geom_dl(aes(label = label), method = list("last.polygons",cex =0.9,fontfamily = "Gill Sans")) + 
  theme(legend.position = "none") +
  scale_colour_brewer(palette="Dark2") +
  ggtitle("Public Insurance")

p_calib_unin <- 
  p_calib_df %>% 
  filter(type=="UNIN") %>% 
  ggplot(aes(x = year, y = value, colour = label)) + 
  geom_point() + 
  # facet_wrap(~type, scales = "free") + 
  theme_minimal() + 
  geom_line() +
  scale_y_continuous(limits = c(0,0.25)) + 
  scale_x_continuous(limits = c(2015,2019)) +
  labs(y = "Fraction of Population Covered", x = "Year") + 
  geom_dl(aes(label = label), method = list("last.polygons",cex =0.9,fontfamily = "Gill Sans")) + 
  theme(legend.position = "none") +
  scale_colour_brewer(palette="Dark2") +
  ggtitle("Uninsured")

p_calib_f <- 
  (p_calib_esi + p_calib_ng) /
  (p_calib_pub + p_calib_unin)
p_calib_f
ggsave(filename = "figures/01-calibration-base-model.png", width = 8, height = 10)

final_probs <- v_calib_map[v_param_names]

R_c <- matrix(nrow = nrow(R), ncol = ncol(R))
R_c[1,1] <- 1 - final_probs["p_ESI_NG"] - final_probs["p_ESI_PUB"]  - final_probs["p_ESI_UNIN"] 
R_c[1,2] <- final_probs["p_ESI_NG"] 
R_c[1,3] <- final_probs["p_ESI_PUB"] 
R_c[1,4] <- final_probs["p_ESI_UNIN"] 

R_c[2,2] <- 1 - final_probs["p_NG_ESI"] - final_probs["p_NG_PUB"]  - final_probs["p_NG_UNIN"] 
R_c[2,1] <- final_probs["p_NG_ESI"] 
R_c[2,3] <- final_probs["p_NG_PUB"] 
R_c[2,4] <- final_probs["p_NG_UNIN"] 

R_c[3,3] <- 1 - final_probs["p_PUB_ESI"] - final_probs["p_PUB_NG"]  - final_probs["p_PUB_UNIN"] 
R_c[3,1] <- final_probs["p_PUB_ESI"] 
R_c[3,2] <- final_probs["p_PUB_NG"] 
R_c[3,4] <- final_probs["p_PUB_UNIN"] 

R_c[4,4] <- 1 - final_probs["p_UNIN_ESI"] - final_probs["p_UNIN_NG"]  - final_probs["p_UNIN_PUB"] 
R_c[4,1] <- final_probs["p_UNIN_ESI"] 
R_c[4,2] <- final_probs["p_UNIN_NG"] 
R_c[4,3] <- final_probs["p_UNIN_PUB"] 

calibration_results <- 
  list(fit = m_calib_res,
       R_c = R_c,
       target = lst_targets,
       test = run_insurance_markov(v_calib_map[v_param_names]))

calibration_results %>%  
  write_rds(here("output/model-calibration-to-acs/01_calibration-results.rds"))

```
