==> R/02_replicate-finkelstein-shepard-hendren.Rmd <==

```{r, echo = FALSE}
knitr::opts_chunk$set(echo = FALSE)
knitr::opts_chunk$set(warning = FALSE)
knitr::opts_chunk$set(message = FALSE)
knitr::opts_chunk$set(cache = TRUE)
```

```{r}
suppressWarnings(suppressMessages(source(here::here("./R/manifest.R"))))
library(directlabels)
library(gt)
source(here("./R/functions_finkelstein-hendren-shepard-replication.R"))
conflict_prefer("set_names","rlang")
conflict_prefer("filter","dplyr")
```

```{r}
# # Import datasets
# 1% Collapsed 2011 Dataset used to replicate the primary RD regressions
df <- 
  haven::read_dta(here("./data/AER_data_files/Data/Collapsed_CommCare/model_analysis_data_by1pctFPL_matchFPL_scaledACS.dta")) %>% 
  mutate(inRD_sample = ifelse((pyear == 2011 & (income_grp >= 135 & income_grp <= 300) & region == ""), 1, 0)) %>%
  mutate(aboveYCut1 = ifelse(income_grp > 150, 1, 0), 
         aboveYCut2 = ifelse(income_grp > 200, 1, 0),
         aboveYCut3 = ifelse(income_grp > 250, 1, 0)) %>%
  mutate(fpl_X_aboveYCut1 = income_grp * aboveYCut1, 
         fpl_X_aboveYCut2 = income_grp * aboveYCut2,
         fpl_X_aboveYCut3 = income_grp * aboveYCut3) %>%
  mutate(fpl_relcut1 = (income_grp - 150), 
         fpl_relcut2 = (income_grp - 200), 
         fpl_relcut3 = (income_grp - 250)) %>%
  mutate(fpl_X_aboveCut1 = fpl_relcut1*aboveYCut1, 
         fpl_X_aboveCut2 = fpl_relcut2*aboveYCut2, 
         fpl_X_aboveCut3 = fpl_relcut3*aboveYCut3) %>%
  mutate(cost_ins_all = cost_ins, 
         cost_ins_H = cost_ins_Hplan, 
         cost_ins_L = cost_ins_Lplan) %>% 
  select(nTotEnr, nTotEnr_Hplan, nNewEnr_new, nNewEnr_new_Hplan, 
         nTotEnr_Lplan,share_Ins = share_Insured, share_H = share_NonCeltPlan, share_CeltiCare,
         aboveYCut1, aboveYCut2, aboveYCut3, income_grp, fpl_X_aboveYCut1, fpl_X_aboveYCut2, fpl_X_aboveYCut3, 
         cost_ins_all, cost_ins_H, cost_ins_L,
         everything()) 


```

```{r}
#########################
# Willingness to Pay RDs
#########################

# Fit Low- Generosity/Premium Plan RD
fit_all <- 
  df %>% 
  estimate_fsh_rd(yvar = "share_Ins")

# Fit High-Low Difference RD
fit_HL <- 
  df %>% 
  estimate_fsh_rd(yvar = "share_H")

#####################
# Cost RDs
#####################

fit_cost_all <- 
  df %>% 
  estimate_fsh_rd(yvar = "cost_ins_all")

fit_cost_L <- 
  df %>% 
  estimate_fsh_rd(yvar = "cost_ins_L")

fit_cost_H <- 
  df %>% 
  estimate_fsh_rd(yvar = "cost_ins_H")
```

```{r}
###############
# Estimate WTP
###############
est_wtp <- 
  get_WTP(coef = fit_all$coef,
             coef_H = fit_HL$coef)
#################
# Estimate Costs
#################
est_cost <- 
  get_cost(coef_cost =  fit_cost_H$coef,
          coef_D  = fit_HL$coef)
```

```{r}
# Extrapolated WTP
df_est_wtp <- 
  est_wtp %>% 
  select(fpl_group,demand_L_ext,demand_H_ext) %>% 
         unnest() %>% 
         select(fpl_group,s,price_L = price, price_H = price1) %>% 
         gather(type,price,-fpl_group,-s) %>% 
         mutate(type = gsub("price_","",type)) %>% 
  select(fpl_group , type, s , wtp = price)

# Extrapolated Costs
df_est_cost <- 
  est_cost %>% 
  pluck("cost_H_ext") %>% 
  set_names(c("fpl150","fpl200","fpl250")) %>% 
  bind_rows(.id  = "fpl_group") %>% 
  select(fpl_group, s = share, cost,avg_cost = premium) %>% 
  mutate(type = "H") %>% 
  na.omit() %>% 
  select(fpl_group, type, s, cost, avg_cost)

df_est <- 
  df_est_wtp %>% 
  left_join(df_est_cost,c("fpl_group","type","s"))

df_est %>% write_rds(here("./output/demand-estimates/01_wtp-and-cost-extrapolated-points.rds"))
```

```{r}
df_est_wtp_points <- 
  est_wtp %>% 
  select(fpl_group,demand_L_est) %>% 
         unnest() %>% 
         select(fpl_group,s = demand_L, price = price) %>% 
  mutate(type = "L") %>% 
  bind_rows(
    est_wtp %>% 
    select(fpl_group,demand_H_est) %>% 
         unnest() %>% 
         select(fpl_group, s= s, price = price) %>% 
      mutate(type = "H")
  ) %>% 
  select(fpl_group , type, s , wtp = price)

df_est_cost_points <-
  est_cost %>% 
    pluck("cost_H_est") %>% 
    set_names(c("fpl150","fpl200","fpl250")) %>% 
    bind_rows(.id  = "fpl_group") %>% 
    mutate(type = "H") %>% 
    select(fpl_group, type, s = share, avg_cost = premium)  %>% 
  bind_rows(
  est_cost %>% 
    pluck("cost_H_est2") %>% 
    set_names(c("fpl150","fpl200","fpl250")) %>% 
    bind_rows(.id  = "fpl_group") %>% 
    mutate(type = "H") %>% 
    select(fpl_group, type, s = share, cost) 
  )

df_est_points <- 
  df_est_wtp_points %>% 
  bind_rows(df_est_cost_points)

df_est_points %>% write_rds(here("./output/demand-estimates/01_wtp-and-cost-points.rds"))
```

```{r}
p <- df_est %>% 
  filter(fpl_group=="fpl150") %>% 
  gather(key,value,-fpl_group,-type,-s) %>% 
  filter(type=="H") %>% 
  ggplot(aes(x = s , y = value,colour = key)) + 
  #facet_wrap(fpl_group~type) + 
  #geom_line(lty=3) + 
  ggthemes::theme_clean() +  
  ylim(c(0,500)) + 
  xlim(c(0,1)) +
  geom_point(data = 
               df_est_points %>% 
               filter(fpl_group=="fpl150") %>% 
               gather(key,value,-fpl_group,-type,-s) %>% 
               filter(type=="H") 
               ) +
  geom_line() +
  geom_line(data = 
               df_est_points %>% 
              filter(fpl_group=="fpl150") %>% 
               gather(key,value,-fpl_group,-type,-s) %>% 
               filter(type=="H") 
               ) + 
  ylab("$ per month") + 
  xlab("Fraction in plan (=s)") #+ 
  #ggtitle("Replication of Finkelstein, Hendren and Shepard (2019)\nFigure 12")

```


```{r, eval = TRUE}
# WTP and Costs (Primary Estimates)
df_wtp_and_costs <- 
  fit_wtp_and_costs(
    coef_wtp_l = fit_all$coef,
    coef_wtp_hl = fit_HL$coef,
    coef_cost_h = fit_cost_H$coef, 
    coef_cost_demand = fit_HL$coef
  ) %>% 
  mutate(iteration = "0")

df_wtp_and_costs %>% 
   write_rds(here("./output/demand-estimates/01_wtp-and-cost.rds"))

df_wtp_and_costs_data <- 
  fit_wtp_and_costs(
    coef_wtp_l = fit_all$coef,
    coef_wtp_hl = fit_HL$coef,
    coef_cost_h = fit_cost_H$coef, 
    coef_cost_demand = fit_HL$coef,
    output_type = "data"
  ) %>% 
  mutate(iteration = "0")


fit_uncertainty  = FALSE

if (fit_uncertainty) {
  df_uncert_wtp_and_costs <-
    1:1000 %>%
    map(~(
      fit_wtp_and_costs(
          coef_wtp_l = MASS::mvrnorm(n = 1, mu = coefficients(fit_all$fit), Sigma = fit_all$vcov),
          coef_wtp_hl = MASS::mvrnorm(n = 1, mu = coefficients(fit_HL$fit), Sigma = fit_HL$vcov),
          coef_cost_h = MASS::mvrnorm(n = 1, mu = coefficients(fit_cost_H$fit), Sigma =fit_cost_H$vcov),
          coef_cost_demand = MASS::mvrnorm(n = 1, mu = coefficients(fit_HL$fit), Sigma = fit_HL$vcov))
    )) %>%
      bind_rows(.id = "iteration")
  
  df_uncert_wtp_and_costs %>% 
   write_rds(here("./output/demand-estimates/01_wtp-and-cost_uncertainty.rds"))
  
   df_uncert_wtp_and_costs_data <-
    1:1000 %>%
    map(~(
      fit_wtp_and_costs(
          coef_wtp_l = MASS::mvrnorm(n = 1, mu = coefficients(fit_all$fit), Sigma = fit_all$vcov),
          coef_wtp_hl = MASS::mvrnorm(n = 1, mu = coefficients(fit_HL$fit), Sigma = fit_HL$vcov),
          coef_cost_h = MASS::mvrnorm(n = 1, mu = coefficients(fit_cost_H$fit), Sigma =fit_cost_H$vcov),
          coef_cost_demand = MASS::mvrnorm(n = 1, mu = coefficients(fit_HL$fit), Sigma = fit_HL$vcov),
      output_type = "data")
    )) %>%
      bind_rows(.id = "iteration")
   
  df_uncert_wtp_and_costs_data %>% 
   write_rds(here("./output/demand-estimates/02_wtp-and-cost_uncertainty-data.rds"))
  
   
}

df_uncert_wtp_and_costs <- read_rds(here("./output/demand-estimates/01_wtp-and-cost_uncertainty.rds"))
df_uncert_wtp_and_costs_data <- read_rds(here("./output/demand-estimates/02_wtp-and-cost_uncertainty-data.rds"))
```

```{r validation}
df_validation <- 
  data.frame(fpl = rep(c(150,150,150,150,150),1), type = "H", prem = c(0,36,105,246,300))  %>% 
    inner_join(df_wtp_and_costs,c("fpl","type")) %>% 
    mutate(estimate = intercept + wtp * prem + i_wtp_2 * prem^2 + i_wtp_3 * prem ^3) %>% 
    mutate(fpl_group = paste0("fpl",fpl)) %>% 
    select(fpl_group,type,wtp = prem, outcome,estimate,iteration) %>% 
    spread(outcome,estimate) %>% 
    gather(key,value,-fpl_group,-type,-s,-iteration)

df_validation_uncert <- 
  data.frame(fpl = rep(c(150,150,150,150,150),1), type = "H", prem = c(0,36,105,246,300)) %>% 
    inner_join(df_uncert_wtp_and_costs,c("fpl","type")) %>% 
    mutate(estimate = intercept + wtp * prem + i_wtp_2 * prem^2 + i_wtp_3 * prem ^3) %>% 
    mutate(fpl_group = paste0("fpl",fpl)) %>% 
    select(iteration,fpl_group,type,wtp = prem, outcome,estimate) %>% 
    spread(outcome,estimate) %>% 
    gather(key,value,-fpl_group,-type,-s,-iteration) %>% 
  tbl_df()

df_validation_uncert_quant <- 
  df_validation_uncert %>% 
  filter(value %in% c(0,36,105,246,300)) %>% 
  ungroup() %>% 
  group_by(fpl_group,type,key,value) %>% 
  summarise_at(vars(s),list(q25 = ~pmax(0,pmin(1,quantile(.,probs = 0.1,na.rm=TRUE))),
                                q75 = ~pmax(0,pmin(1,quantile(.,probs = 0.9, na.rm=TRUE)))))

# p2 <- 
#   p + 
#   geom_point(data = df_validation_uncert %>% filter(fpl_group=="fpl150"),alpha = 0.05)  +
#     geom_point(data = df_validation %>% filter(fpl_group=="fpl150"),colour = "black",pch = 3)  + 
#   theme(text = element_text(family = "Gill Sans"),
#         plot.title = element_text(family = "Gill Sans", face = "plain"),
#         plot.background = element_blank()) + 
#   geom_dl(aes(label = key), method = list("last.polygons",cex =0.9,fontfamily = "Gill Sans"))  + 
#   scale_x_continuous(limits = c(0,1.3),breaks = seq(0,1,0.2)) + 
#   theme(legend.position="none")
 
p2 <- 
  p + 
  geom_line(
    data = 
        df_uncert_wtp_and_costs_data %>% 
        filter(fpl_group=="fpl150" & type == "H") %>% 
        gather(key,value,-iteration,-fpl_group,-type,-s) %>% 
        filter(key=="wtp"),
    aes(group = iteration) ,
    alpha = 0.02
  ) + 
  geom_line(
    data = 
      df_uncert_wtp_and_costs_data %>% 
      filter(fpl_group=="fpl150" & type == "H") %>% 
      gather(key,value,-iteration,-fpl_group,-type,-s) %>% 
      filter(key=="cost"),
    aes(group = iteration) ,
    alpha = 0.02
  ) + 
  geom_line(
    data = 
      df_uncert_wtp_and_costs_data %>% 
      filter(fpl_group=="fpl150" & type == "H") %>% 
      gather(key,value,-iteration,-fpl_group,-type,-s) %>% 
      filter(key=="avg_cost"),
    aes(group = iteration) ,
    alpha = 0.02
  ) + 
  theme(text = element_text(family = "Gill Sans"),
        plot.title = element_text(family = "Gill Sans", face = "plain"),
        plot.background = element_blank()) +
  geom_dl(aes(label = key), method = list("last.polygons",cex =0.9,fontfamily = "Gill Sans"))  +
  scale_x_continuous(limits = c(0,1.1),breaks = seq(0,1,0.2)) +
  theme(legend.position="none") 

p_uncert <- df_est %>% 
  filter(fpl_group=="fpl150") %>% 
  gather(key,value,-fpl_group,-type,-s) %>% 
  filter(type=="H") %>% 
  ggplot(aes(x = s , colour = key)) + 
  #facet_wrap(fpl_group~type) + 
  #geom_line(lty=3) + 
  ggthemes::theme_clean() +  
  #ylim(c(0,2000)) + 
  xlim(c(0,1)) +
  geom_point(data = 
               df_est_points %>% 
               filter(fpl_group=="fpl150") %>% 
               gather(key,value,-fpl_group,-type,-s) %>% 
               filter(type=="H") , 
             aes(y = value)
  ) +
  geom_line(aes(y = value)) +
  geom_line(data = 
              df_est_points %>% 
              filter(fpl_group=="fpl150") %>% 
              gather(key,value,-fpl_group,-type,-s) %>% 
              filter(type=="H") , 
            aes(y = value)
  ) + 
  ylab("$ per month") + 
  xlab("Fraction in plan (=s)") + 
  geom_ribbon(data = 
                df_uncert_wtp_and_costs_data %>% 
                filter(fpl_group=="fpl150" & type == "H") %>% 
                gather(key,value,-iteration,-fpl_group,-type,-s) %>% 
                filter(key=="wtp") %>% 
                group_by(s) %>% 
                summarise_at(vars(value),list(low = ~quantile(.,0.025,na.rm=TRUE), high = ~quantile(.,0.975,na.rm=TRUE))) %>% 
                ungroup() %>% 
                mutate(key = "wtp"), 
              aes(x = s, ymin = low, ymax = high), alpha = 0.1, fill = "blue", lty=0) + 
  geom_ribbon(data = 
                df_uncert_wtp_and_costs_data %>% 
                filter(fpl_group=="fpl150" & type == "H") %>% 
                gather(key,value,-iteration,-fpl_group,-type,-s) %>% 
                filter(key=="cost")%>% 
                group_by(s) %>% 
                summarise_at(vars(value),list(low = ~quantile(.,0.025,na.rm=TRUE), high = ~quantile(.,0.975,na.rm=TRUE))) %>% 
                ungroup() %>% 
                mutate(key = "cost"), 
              aes(x = s, ymin = pmax(0,low), ymax = high), alpha = 0.1, fill = "green",lty=0) +
  geom_ribbon(data = 
                df_uncert_wtp_and_costs_data %>% 
                filter(fpl_group=="fpl150" & type == "H") %>% 
                gather(key,value,-iteration,-fpl_group,-type,-s) %>% 
                filter(key=="avg_cost")%>% 
                group_by(s) %>% 
                summarise_at(vars(value),list(low = ~quantile(.,0.025,na.rm=TRUE), high = ~quantile(.,0.975,na.rm=TRUE))) %>% 
                ungroup() %>% 
                mutate(key = "avg_cost"), 
              aes(x = s, ymin = pmax(0,low), ymax = high), alpha = 0.1, fill = "red",lty = 0) +
   geom_dl(aes(label = key,y = value), method = list("last.polygons",cex =0.9,fontfamily = "Gill Sans")) +
  scale_x_continuous(limits = c(0,1.1),breaks = seq(0,1,0.2)) +
  theme(legend.position="none") 

ggsave(p_uncert,filename = here("./figures/01_FHS-replication-fig12.png"), width = 8, height = 6)

####################
# For Presentation
####################
df_subsidy25 <- 
  data.frame(fpl = rep(c(150),1), type = "H", prem = c(36))  %>% 
    inner_join(df_wtp_and_costs,c("fpl","type")) %>% 
    mutate(estimate = intercept + wtp * prem + i_wtp_2 * prem^2 + i_wtp_3 * prem ^3) %>% 
    mutate(fpl_group = paste0("fpl",fpl)) %>% 
    select(fpl_group,type,wtp = prem, outcome,estimate,iteration) %>% 
    spread(outcome,estimate) %>% 
    gather(key,value,-fpl_group,-type,-s,-iteration)


p_prez <- 
  p + 
   geom_point(data = df_subsidy25  %>% filter(fpl_group=="fpl150"),colour = "black",pch = 2)  +
    theme(text = element_text(family = "Gill Sans"),
        plot.title = element_text(family = "Gill Sans", face = "plain"),
        plot.background = element_blank()) +
  geom_dl(aes(label = key), method = list("last.polygons",cex =0.9,fontfamily = "Gill Sans"))  +
  scale_x_continuous(limits = c(0,1.1),breaks = seq(0,1,0.2)) +
  theme(legend.position="none") + 
  geom_hline(aes(yintercept = 36), lty=2) +
  scale_x_continuous(limits = c(0,1.1),breaks = sort(c(seq(0,1,0.2),0.85)))
ggsave(p_prez , filename = here("./figures/01_FHS-replication-fig12-for-presentation.png"))

```
\newpage 
```{r}
validation_summ <- 
  df_validation_uncert %>% 
  tbl_df() %>% 
  spread(key,value) %>% 
  group_by(wtp) %>% 
  summarise_at(vars(s,cost),list(sd = ~sd(pmax(0,.))))
df_validation %>% 
  spread(key,value) %>% 
  left_join(validation_summ,"wtp") %>% 
  select(premium = wtp, takeup= s, takeup_se = s_sd, cost = cost, cost_se = cost_sd) %>% 
  gt() %>% 
  # tab_header(
  #   title = "Replication of Finkelstein, Hendren and Shepard (2019) Table 2"
  # ) %>% 
  fmt_number(
    columns = vars(takeup),
    decimals = 3,
    use_seps = FALSE
  ) %>% 
    fmt_number(
    columns = vars(takeup_se),
    decimals = 2,
    use_seps = FALSE
  ) %>% 
  fmt_currency(
    columns = vars(cost,cost_se),
    decimals = 0,
    use_seps = FALSE
  ) %>% 
  gtsave(filename = here("./figures/01_FHS-replication-table2.png"))
```



```{r, eval = TRUE}
# get_mvpf_subsidy(wtp = .x$wtp, cost = .x$cost, )

fn_uncomp <- function(cost, uninsured_oop_share , phi ) {
  # x is the share of the uninsured’s total health care costs that they pay out of pocket 
  # φ denotes the percentage increase in costs that result from insurance coverage (moral hazard)
  (1 - uninsured_oop_share) * (cost / (1 + phi))
}

df_mvpf <-
  df_wtp_and_costs_data %>% 
  bind_rows(df_uncert_wtp_and_costs_data) %>% 
  mutate(mvpf_orig = 1 / (1 + ((cost - wtp) / (-s * ((wtp-lag(wtp))/.01))))) %>% 
  filter(type == "H" & fpl_group =="fpl150") %>% 
  na.omit()

get_mvpf_subsidy <- function(df, cost, wtp, s, gov_incidence = 0, uninsured_oop_share = 0.2, phi = 0.25, eta = 0) {
  df %>% 
    #mutate(numerator = 1) %>% 
    #mutate(denominator = (1 + (({{cost}} - {{wtp}}) / (-{{s}} * (({{wtp}}-lag({{wtp}}))/.01))))) %>% 
    mutate(gov_incidence = gov_incidence, 
          x_oop_uninsured = uninsured_oop_share, 
          phi_moral_hazard = phi, 
          eta_incidence_lowinc = eta) %>% 
    mutate(uncomp = fn_uncomp(cost = cost,uninsured_oop_share = uninsured_oop_share, phi = phi)) %>% 
    mutate(
      mvpf_num = {{s}} + 
        eta  * 
          (
            uncomp / 
              (-1 * (( {{wtp}} - lag({{wtp}})) / 0.01))
          )
    ) %>% 
    mutate(
      mvpf_denom =  {{s}} + 
        ((pmax(0,{{cost}} - gov_incidence * uncomp - {{wtp}})) / 
           (-1 * (({{wtp}} - lag({{wtp}}))/.01)))
    ) %>% 
    mutate(mvpf = mvpf_num / mvpf_denom)

}

mvpf_no_uncomp <-   
  df_wtp_and_costs_data %>% 
  bind_rows(df_uncert_wtp_and_costs_data) %>% 
  get_mvpf_subsidy(cost = cost, wtp = wtp , s = s, uninsured_oop_share = 1, phi = 0, gov_incidence = 0, eta = 0) %>% 
  filter(type=="H" & fpl_group=="fpl150" & s %in% c(0.3, 0.9))   %>% 
  mutate(scenario = "No Uncompensated\nCare")
  
# Incidence on Government
mvpf_govt<-   
  df_wtp_and_costs_data %>% 
  bind_rows(df_uncert_wtp_and_costs_data) %>% 
  get_mvpf_subsidy(cost = cost, wtp = wtp , s = s, uninsured_oop_share = 0.2, phi = 0.25, gov_incidence = 1, eta = 0) %>% 
  filter(type=="H" & fpl_group=="fpl150" & s %in% c(0.3, 0.9))  %>% 
  mutate(scenario = "Incidence on\nGovernment")
  
# Incidence on Low-Income Uninsured
mvpf_lowinc1 <-   
  df_wtp_and_costs_data %>% 
  bind_rows(df_uncert_wtp_and_costs_data) %>% 
  get_mvpf_subsidy(cost = cost, wtp = wtp , s = s, uninsured_oop_share = 0.2, phi = 0.25, gov_incidence = 0, eta = 1) %>% 
  filter(type=="H" & fpl_group=="fpl150" & s %in% c(0.3, 0.9))   %>% 
  mutate(scenario = "Incidence on Low\nIncome Uninsured")

# Incidence on Affulent
mvpf_lowinc0_5 <-   
  df_wtp_and_costs_data %>% 
  bind_rows(df_uncert_wtp_and_costs_data) %>% 
  get_mvpf_subsidy(cost = cost, wtp = wtp , s = s, uninsured_oop_share = 0.2, phi = 0.25, gov_incidence = 0, eta = 0.5) %>% 
  filter(type=="H" & fpl_group=="fpl150" & s %in% c(0.3, 0.9))   %>% 
  mutate(scenario = "Incidence on\nAffulent")

df_mvpf_replication <- 
  mvpf_no_uncomp %>% 
    bind_rows(mvpf_govt) %>% 
    bind_rows(mvpf_lowinc1) %>% 
    bind_rows(mvpf_lowinc0_5) %>% 
  mutate(s = paste0(s*100,"% Insured")) %>% 
  mutate(scenario = factor(scenario, levels = levels(factor(scenario))[c(1,3,2,4)]))
                           
sdf_mvpf_replication <- 
  mvpf_no_uncomp %>% 
    bind_rows(mvpf_govt) %>% 
    bind_rows(mvpf_lowinc1) %>% 
    bind_rows(mvpf_lowinc0_5) 

p_tmp <- 
  df_mvpf_replication %>%
  ggplot( aes(x = scenario,fill = s)) +
  scale_x_discrete(limits = rev(levels(df_mvpf_replication$scenario ))) +
  geom_col(data = df_mvpf_replication  %>% filter(iteration == 0), aes(y = mvpf),position = "dodge") +
  ggthemes::theme_clean() +
  geom_hline(aes(yintercept = 0.9),lty=2) +
  geom_hline(aes(yintercept = 0.2),lty=2) +
  scale_y_continuous(limits = c(0,1.5), breaks = seq(0,1.5,0.25)) +
  geom_text(data = df_mvpf_replication  %>% filter(iteration == 0),
            aes(y = mvpf+.08,label = round(mvpf,2)), position = position_dodge(0.9), family = "Gill Sans") +
  theme(legend.position = "bottom") +
  #ggtitle("Replication of Finkelstein, Hendren and Shepard (2017)\nFigure 14") +
  ylab("MVPF")  + xlab("") +
  theme(text = element_text(family = "Gill Sans"),
        plot.title = element_text(family = "Gill Sans", face = "plain"),
        plot.background = element_blank(), legend.title = element_blank()) 

p_tmp %>% ggsave(filename = here("./figures/01_FHS-replication-fig14.png"))
```
