==> R/03_simulate-subsidy-and-medicaid.Rmd <==
---
output: html_document
editor_options: 
  chunk_output_type: console
---

```{r, echo = FALSE}
knitr::opts_chunk$set(echo = FALSE)
knitr::opts_chunk$set(warning = FALSE)
knitr::opts_chunk$set(message = FALSE)
knitr::opts_chunk$set(cache = TRUE)
```

```{r}
# Math characters in ggplot
# https://rstudio-pubs-static.s3.amazonaws.com/136237_170402e5f0b54561bf7605bdea98267a.html
# https://en.wikibooks.org/wiki/Unicode/Character_reference/0000-0FFF

suppressWarnings(suppressMessages(source(here::here("./R/manifest.R"))))
wrap_string <- function(x,width) paste0(strwrap(x,width = width),collapse = "\n")

results <- 
  read_rds(here("output/dd-estimates/primary-results.rds"))
dd_lm <- results$dd_lm
dd_lm_ri <- results$dd_lm_ri

#!!! This file is constructed in 01_construct-and-calibrate-model.Rmd 
df_calib <- 
  read_rds(here("output/model-calibration-to-acs/01_calibration-results.rds"))

#!!! These files are written in 02_replicate-finkelstein-shepard-hendren.Rmd
df_wtp_and_costs <-  read_rds(here("./output/demand-estimates/01_wtp-and-cost.rds"))
df_uncert_wtp_and_costs <- read_rds(here("./output/demand-estimates/01_wtp-and-cost_uncertainty.rds"))
df_uncert_wtp_and_costs_data <- read_rds(here("./output/demand-estimates/02_wtp-and-cost_uncertainty-data.rds"))
```

```{r}
##################
# Set Parameters
##################

  # estimated in in 01_simulation_fit-medciaid-DD.Rmd

  # Ex Ante Occupancy Vector
  p <- 
    dd_lm %>% 
    filter(!is.na(ex_ante)) %>% 
    select(ex_ante,starts_with("p_Z1_T0"), starts_with("p_Z0_T0")) %>% 
    na.omit() %>% 
    tbl_df()  %>% 
    # Need to get an overall distribution here, but using the 
    # nontreated states for now.
    pull(p_Z0_T0) %>% 
    as.list() %>% 
    set_names(paste0("p_",insurance_sipp_lut))
   
  # Baseline Transition Matrix
  df_R <- 
    dd_lm %>% 
    select(ex_ante,starts_with("R0")) %>% 
    na.omit() %>% 
    gather(estimate, value, -ex_ante) %>% 
    mutate(ex_post = estimate) %>% 
    mutate(ex_post = gsub("R0_","",ex_post)) %>% 
    mutate(param = paste0("R_",ex_ante,"_TO_",ex_post)) %>% 
    select(param,value) 
  R <- df_R$value %>% set_names(df_R$param)

  # DD Estimate for Medicaid Expansion
  
  df_DD <- 
    dd_lm %>% 
    select(ex_ante,starts_with("DD")) %>% 
    na.omit() %>% 
    gather(estimate, value, -ex_ante) %>% 
    mutate(ex_post = estimate) %>% 
    mutate(ex_post = gsub("DD_","",ex_post)) %>% 
    mutate(param = paste0("DD_",ex_ante,"_TO_",ex_post)) %>% 
    select(param,value)
  DD <- df_DD$value %>% set_names(df_DD$param)
  
# Define Parameter List object
params <- 
  c(p,R,DD)

## Tables for Publication

  R <-  params[grep("^R_",names(params))] %>% unlist() %>% 
    data.frame() %>% 
    rownames_to_column(var = "type") %>% 
    separate(type,into= c("exa","exp"), sep ="_TO_") %>% 
    set_names(c("exa","exp","value")) %>% 
    spread(exp,value) %>% 
    select(-exa) %>% 
    as.matrix()
  
  DD <-  params[grep("^DD_",names(params))] %>% unlist() %>% 
    data.frame() %>% 
    rownames_to_column(var = "type") %>% 
    separate(type,into= c("exa","exp"), sep ="_TO_") %>% 
    set_names(c("exa","exp","value")) %>% 
    spread(exp,value) %>% 
    select(-exa) %>% 
    as.matrix()
  
#####################
# MEDICAID
#####################
params$G_Cx <- 2721 # Medicaid Spending among control compliers
params$G <- 3600 # Medicaid spending among treatment compliers
params$OOP_Cx <- 569 # OOP spending among control compliers
params$OOP_Tx <- 0 # OOP spending among treatment compliers
params$sigma <- 3 # Coefficient of absolute risk aversion (needed only if we model I later)
params$I <- 760 # pure insurance component -- note this can be further estimated.
params$gov_incidence = 0.5
params$fudge <-0
params$v_i <- 0.65
params$v_j <- 1.15


###########
# SUBSIDY
###########
params$pop_fpl = 150
params$plan_type = "H"
params$uninsured_oop_share = params$OOP_Cx / params$G_Cx
params$phi = 0.25
params$eta = 0.75
params$plan_premium = 25
params$welfare_weight = 0.5
params$frac_uninsured_elig = 0.7

# Parameters that are functions of other parameters
params$takeup_coef = get_takeup_coef(params = params)
params$cost_coef = get_cost_coef(params = params)

subsidy_test <- simulate_subsidy(params)
medicaid_test <- calculate_wtp_public(params)
```

```{r}
dist_lut <- c("qnorm" = "Normal",
              "qunif" = "Uniform",
              "qnorm01" = "Normal")


params_lut <- 
  c("G" = "Average cost to the government per Medicaid recipient",
    "G_Cx" = "Medicaid spending among control compliers",
    "C_H" = "Cost of marginal enrollee in private plan",
    "OOP_Cx" = "OOP spending among control compliers",
    "OOP_Tx" = "OOP spending among treatment compliers",
    "sigma" = "Coefficient of absolute risk aversion (NOT CURRENTLY USED)",
    "I" = "Pure insurance value of Medicaid",
    "gov_incidence" = "Government incidence of uncompensated care",
    "pop_fpl" = "FPL of recipients",
    "uninsured_oop_share" = "Out-of-pocket share of expenditures among uninsured",
    "phi" = "Moral hazard effect of insurance",
    "eta" = "Social welfare weight",
    "plan_premium" = "Subsidized monthly premium",
    "subsidy_takeup" = "Subsidy take-up",
    "v_i" = "Welfare weight for targeted income group",
    "v_j" = "Welfare weight for high-income group",
    "takeup" = "Fraction taking up subsidized coverage",
    "frac_uninsured_elig" = "Fraction of uninsured population eligible for subsidies")

params_lut_wrapped <- lapply(params_lut,function(x) wrap_string(x,width = 50))
```

```{r}
# Get estimate of takeup uncertainty (SD of RI sampling distribution)
  p_se <- 
    dd_lm_ri %>% 
    group_by(replicate) %>% 
    nest() %>% 
    mutate(p = map(data,~(
      .x %>% 
        filter(!is.na(ex_ante)) %>% 
        select(ex_ante,starts_with("p_Z1_T0"), starts_with("p_Z0_T0")) %>% 
        na.omit() %>% 
        tbl_df()  %>% 
        # Need to get an overall distribution here, but using the 
        # nontreated states for now.
        select(value = p_Z0_T0) %>% 
        mutate(param = insurance_sipp_lut)
    ))) %>% 
    select(-data) %>% 
    unnest() %>% 
      group_by(param) %>% 
      summarize(value = sd(value)) %>% 
    pull(value) %>% 
    as.list() %>% 
    set_names(paste0("p_",insurance_sipp_lut))
    
  df_R_se <- 
    dd_lm_ri %>% 
    group_by(replicate) %>% 
    nest() %>% 
    mutate(p = map(data,~(
      .x %>% 
        select(ex_ante,starts_with("R0")) %>% 
        na.omit() %>% 
        gather(estimate, value, -ex_ante) %>% 
        mutate(ex_post = estimate) %>% 
        mutate(ex_post = gsub("R0_","",ex_post)) %>% 
        mutate(param = paste0("R_",ex_ante,"_TO_",ex_post)) %>% 
        select(param,value)
    ))) %>% 
    select(-data) %>% 
    unnest() %>% 
    group_by(param) %>% 
    summarize(value = sd(value)) 
  R_se <- df_R_se$value %>% set_names(df_R_se$param)
  R_se <- as.list(R_se)
  df_DD_se <- 
    dd_lm_ri %>% 
    group_by(replicate) %>% 
    nest() %>% 
    mutate(p = map(data,~(
      .x %>% 
        select(ex_ante,starts_with("DD")) %>% 
        na.omit() %>% 
        gather(estimate, value, -ex_ante) %>% 
        mutate(ex_post = estimate) %>% 
        mutate(ex_post = gsub("DD_","",ex_post)) %>% 
        mutate(param = paste0("DD_",ex_ante,"_TO_",ex_post)) %>% 
        select(param,value)
    ))) %>% 
    select(-data) %>% 
    unnest() %>% 
    group_by(param) %>% 
    summarize(value = sd(value)) 
  DD_se <- df_DD_se$value %>% set_names(df_DD_se$param)
  DD_se <- as.list(DD_se)
```


```{r}

qnorm01 <- function(x,...) pmax(0,pmin(1,qnorm(p = x,...)))
### Probabilistic Sensitivity Analysis
params_psa_vals <- 
  
params_psa <- params
params_psa$gov_incidence = function(x) qunif(p = x, min = 0, max = 1) # Incidence on Government
params_psa$eta = function(x) qunif(p = x, min =0.5, max = 1) # Marginal utility of 
params_psa$v_i = function(x) qnorm(p = x, mean = params$v_i,sd = 0.02) # Welfare Weightfor targeted group
params_psa$v_j = function(x) qnorm(p = x, mean = params$v_j,sd = 0.02) # Social Welfare for high-income group
params_psa$phi = function(x) qnorm(p = x, mean = 0.25,sd = 0.05) # Moral Hazard Effects of Insurance
params_psa$uninsured_oop_share =  function(x) qnorm(p = x, mean = 0.2, sd = 0.1) #function(x) pmax(0,pmin(1,qnorm(p = x, mean = 0.2,sd = 0.1))) # Share OOP for Uninsured
params_psa$G_Cx =  function(x) qnorm(p = x, mean = params$G_Cx, sd = 20) # Medicaid Spending among control compliers
params_psa$G = function(x) qnorm(p = x, mean = params$G, sd = 20) # Medicaid spending among treatment compliers
params_psa$OOP_Cx = function(x) qnorm(p = x, mean = params$OOP_Cx, sd = 50)  # OOP spending among control compliers
params_psa$OOP_Tx = function(x) qnorm(p = x, mean = params$OOP_Tx, sd = 0) # OOP spending among treatment compliers
params_psa$sigma = function(x) qnorm(p = x, mean =  params$sigma, sd = 0) # Coefficient of absolute risk aversion (needed only if we model I later)
params_psa$I = function(x) qnorm(p = x, mean =  params$I, sd = 200) # pure insurance component -- note this can be further estimated.
params_psa$fudge = function(x) qnorm(p = x, mean =  params$fudge, sd = 0) # pure insurance component -- note this can be further estimated.
params_psa$gov_incidence = function(x) qunif(p = x, min = 0, max = 1)
params_psa$frac_uninsured_elig = function(x) qnorm(p = x, mean =  params$frac_uninsured_elig, sd = .05)
#params_psa$welf_weight = function(x) qunif(p = x, min = 0, max = 1)
    
# Only use Halton draws for PSA parameters that aren't fixed.
which_are_fns <- params_psa %>% map_lgl(~is.function(.x)) 

```

```{r}
get_psa_values <- function(x) {

  foo <- deparse(x)
  parameter_name <- str_split(gsub("list\\(","",foo[1])," = ")[[1]][1]
  dist = dist_lut[str_split(foo[2],"\\(")[[1]][1]]
  
    gsub(" ","",str_split(str_split(foo[2],"\\(|\\)")[[1]][2],",")[[1]]) %>% data.frame() %>% 
    set_names("parameter") %>% 
    separate(parameter,"=",into = c("parameter","value")) %>% 
    mutate(from_params_list = as.integer(grepl("params",value))) %>% 
    mutate(value = gsub("params\\$","",value))  %>% 
    mutate(rowname = parameter_name) %>% 
    filter(value != "x")  %>% 
    mutate(value = ifelse(from_params_list==1,params[value],value)) %>% 
    select(rowname,parameter,value) %>% 
    spread(parameter,value) %>% 
      mutate(distribution = dist) %>% 
    select(-rowname) %>% 
      select(distribution,everything()) %>% 
      tbl_df() %>% 
      mutate_at(vars(-distribution),list(~paste0(unlist(.))))
}


psa_values <- 
  params_psa[which_are_fns] %>% 
  map(~(
  get_psa_values(.x)
))  %>% bind_rows(.id = "rowname")

tbl <- 
  params %>% unlist() %>% data.frame() %>% 
  rownames_to_column() %>% 
  filter(rowname %in% names(params_lut)) %>% 
  mutate(desc = as.character(params_lut[rowname])) %>% 
  select(3,1,2) %>% 
  left_join(psa_values,c("rowname")) %>% 
  mutate_at(vars(3),function(x) as.numeric(paste0(x))) %>% 
  mutate_at(vars(3),function(x) ifelse(x>1000,x/1000,ifelse(x<=1,round(x,3), round(x,1)))) %>%
  filter(rowname!="sigma")
  #mutate(desc = gsub("$","\\$",desc)) %>% 
  #set_colnames(c("Parameter Description","Short Name","Value")) 
  #tangram(., as.character=TRUE, style="lancet", caption="Model Parameters",id = "params")

p_tbl <- 
tbl %>% 
  #filter(!grepl("NOT CURRENTLY",desc)) %>% 
  gt(rowname_col = "Parameter Description") %>% 
  tab_header(
    title = "Model Parameters"
  ) %>% 
  tab_options(
    table.width = pct(75)
  ) %>% 
  fmt_missing(columns = 3:8, missing_text = "") %>% 
  cols_label(. = "Value", 
             desc = "Parameter",
             rowname = "Short Name",
             distribution = "PSA Distribution",
             mean = "Mean",
             sd = "Standard Deviation",
             max = "Maximum",
             min = "Minimum") %>% 
  tab_row_group(group = "Shared Parameters", rows = which(tbl$rowname %in% c("gov_incidence","v_i","v_j","pop_fpl","uninsured_oop_share"))) %>% 
  tab_row_group(group = "Medicaid Parameters", rows = which(tbl$rowname %in% c("G_Cx","G","OOP_Cx","OOP_Tx","I"))) %>% 
  tab_row_group(group = "Price-Fixed Subsidy Parameters", rows = which(tbl$rowname %in% c("plan_premium","frac_uninsured_elig","phi","eta")))

p_tbl 
gtsave(p_tbl, filename = here("./figures/03_simulate-subsidy-baseline-parameters.png"))
```


```{r}
p_tbl_shared <- 
  tbl %>% 
  filter(rowname %in% c("gov_incidence","v_i","v_j","pop_fpl","uninsured_oop_share")) %>% 
  #filter(!grepl("NOT CURRENTLY",desc)) %>% 
  gt(rowname_col = "Parameter Description") %>% 
  tab_header(
    title = "Model Parameters Shared Across Strategies"
  ) %>% 
  tab_options(
    table.width = pct(75)
  ) %>% 
  fmt_missing(columns = 3:8, missing_text = "") %>% 
  cols_label(. = "Value", 
             desc = "Parameter",
             rowname = "Short Name",
             distribution = "PSA Distribution",
             mean = "Mean",
             sd = "Standard Deviation",
             max = "Maximum",
             min = "Minimum") 

gtsave(p_tbl_shared , filename = here("./figures/03_simulate-subsidy-baseline-parameters-shared.png"))


p_tbl_medicaid <- 
  tbl %>% 
  filter(rowname %in% c("G_Cx","G","OOP_Cx","OOP_Tx","I")) %>% 
  #filter(!grepl("NOT CURRENTLY",desc)) %>% 
  gt(rowname_col = "Parameter Description") %>% 
  tab_header(
    title = "Model Parameters: Medicaid Expansion"
  ) %>% 
  tab_options(
    table.width = pct(75)
  ) %>% 
  fmt_missing(columns = 3:8, missing_text = "") %>% 
  cols_label(. = "Value", 
             desc = "Parameter",
             rowname = "Short Name",
             distribution = "PSA Distribution",
             mean = "Mean",
             sd = "Standard Deviation",
             max = "Maximum",
             min = "Minimum") 

gtsave(p_tbl_medicaid , filename = here("./figures/03_simulate-subsidy-baseline-parameters-medicaid.png"))



p_tbl_subsidy<- 
  tbl %>% 
  filter(rowname %in% c("plan_premium","frac_uninsured_elig","phi","eta")) %>% 
  bind_rows(data.frame(desc = "Coverage Takeup: Regression-Discontinuity Estimates", rowname = "beta", distribution="Multivariate Normal"))   %>% 
  #filter(!grepl("NOT CURRENTLY",desc)) %>% 
  gt(rowname_col = "Parameter Description") %>% 
  tab_header(
    title = "Model Parameters: Price-Linked Subsidy"
  ) %>% 
  tab_options(
    table.width = pct(75)
  ) %>% 
  fmt_missing(columns = 3:8, missing_text = "") %>% 
  cols_label(. = "Value", 
             desc = "Parameter",
             rowname = "Short Name",
             distribution = "PSA Distribution",
             mean = "Mean",
             sd = "Standard Deviation",
             max = "Maximum",
             min = "Minimum") 

gtsave(p_tbl_subsidy , filename = here("./figures/03_simulate-subsidy-baseline-parameters-subsidy.png"))
```


```{r}
# Raw halton draw
draws <- randtoolbox::halton(n = 1e3, dim = length(params_psa[which_are_fns])) %>% as.matrix()

set.seed(123)

halton_draws <- 
  draws %>% 
  data.frame() %>% 
  magrittr::set_names(names(params_psa[which_are_fns]))  %>% 
  pmap(list) %>% 
  map(~(get_params(.x,params=params_psa[which_are_fns]))) %>% 
  map(~(data.frame(.x))) %>% bind_rows() %>% 
  pmap(list)

params_halton <- 
  halton_draws %>% 
  map(~(c(.x,params_psa[!which_are_fns]))) 

params_halton_transpose <- 
  params_halton %>% 
  transpose()

params_halton_transpose$takeup_coef <- 
  params_halton %>% 
  map( ~get_takeup_coef(df=df_uncert_wtp_and_costs, params = .x))  
  #map( ~get_takeup_coef(df=df_wtp_and_costs, params = .x))  

params_halton_transpose$cost_coef <- 
  params_halton %>% 
  map( ~get_cost_coef(df=df_uncert_wtp_and_costs, params = .x))  
  #map( ~get_cost_coef(df=df_wtp_and_costs, params = .x))  

params_halton_final <- 
  params_halton_transpose %>% 
  transpose()

# Simulate Subsidy
df_psa <- 
  params_halton_final %>% 
  map(~(simulate_subsidy(.x))) %>% 
  map(~(data.frame(.x))) %>% 
  bind_rows(.id = "iteration")

df_psa_medicaid <- 
  params_halton_final %>% 
  map(~(calculate_wtp_public(.x))) %>% 
  map(~(data.frame(.x))) %>% 
  bind_rows(.id = "iteration") %>% 
  select(iteration, mvpf, mvpf_num, mvpf_denom) %>% 
  mutate(strategy = "medicaid")

df_est <- 
  simulate_subsidy(params) %>% 
  data.frame() %>% 
  mutate(iteration = "0") %>% 
  mutate(label_reformed = paste0(" Modeled Estimate:\n $",params$plan_premium,"/mo premium\n (MVPF = ",round(mvpf,3),")")) 

df_est_medicaid <- 
  calculate_wtp_public(params) %>% 
  data.frame() %>% 
  mutate(iteration = "0") %>% 
  select(iteration, mvpf, mvpf_num, mvpf_denom) %>% 
  mutate(strategy = "medicaid") %>% 
  mutate(label_reformed = paste0(" Modeled Estimate:\n Medicaid \n (MVPF = ",round(mvpf,3),")"))  

```

```{r baseline_results}
simulated_policy <- cov_sim(params)

baseline_results <- 
  simulated_policy %>% 
  gather(key,value) %>% 
  filter(grepl("esi_own|priv_oth|public|uninsured|subsidy_mvpf$|med_mvpf$",key))  %>% 
  separate(key,into = c("policy","type1","type2","type3"),"_") %>% 
  mutate(type2 = ifelse(grepl("mvpf",type1),"mvpf",type2)) %>% 
  select(policy,type = type2, value) %>% 
  spread(policy,value) %>% 
  mutate(med = ifelse(type!="mvpf",med - baseline,med)) %>% 
  mutate(subsidy = ifelse(type!="mvpf",subsidy - baseline,subsidy)) %>% 
  mutate(type = factor(type,levels = c("esi","priv","public","uninsured","mvpf"), 
                       labels = c("Employer-Sponsored Insurance","Private Non-Group","Public","Uninsured","Marginal Value of Public Funds"))) %>% 
  arrange(type) %>% 
  gt(rowname_col = "type") %>% 
  tab_stubhead(label = "") %>% 
  cols_label(baseline = "Baseline",
             med = "Expand In-Kind Benefits\n(Medicaid)",
             subsidy = "Price-Linked Subsidy") %>% 
  fmt_number(columns = 1:4, decimals = 3) %>% 
  fmt_missing(columns = 1:4,missing_text="") %>%
  tab_row_group(group = "", rows = c("Marginal Value of Public Funds")) %>% 
  tab_row_group(group = "Coverage Effects", rows = c("Employer-Sponsored Insurance","Private Non-Group","Public","Uninsured"))

baseline_results %>% gtsave(filename = here("figures/99_baseline-results.png"))
```




```{r}
lambda_lut <- c("lambda_1" = "\u3bB = 0.2",
                "lambda_2" = "\u3bB = 0.9",
                "lambda_3" = "\u3bB = 1.2")



df_lambda <- 
  data.frame(
    mvpf_num = seq(0,4,.1)
  ) %>% 
  mutate(lambda_1 = 0.2,
         lambda_2 = 0.9,
         lambda_3 = 1.2) %>% 
  gather(lambda,value,-mvpf_num) %>% 
    mutate(lambda = lambda_lut[lambda]) %>% 
  mutate(mvpf_denom = mvpf_num * 1 / value) %>% 
  filter(mvpf_denom<4)

p_tmp <- 
  df_psa %>%
  ggplot(aes(x = mvpf_num, y = mvpf_denom)) +
  geom_line(data = df_lambda, aes(lty = lambda)) +
  geom_dl(data = df_lambda, aes(label = lambda), method = list("angled.boxes", cex =0.9)) +
  geom_point(alpha = 0.05,pch = 19) +
  ggthemes::theme_clean() +
  scale_y_continuous(limits = c(0,4), breaks = seq(0,4,.25), name = "Cost") +
  scale_x_continuous(limits = c(0,4), breaks = seq(0,4,0.25), name = "Benefit") +
  #geom_abline(intercept = 0, slope = 1 / .9, lty = 2)   +
  #geom_abline(intercept = 0, slope = 1 / .2, lty = 3)   +
  #eom_rug(sides = "tr", length = unit(0.01,"npc"))  +
  geom_point(data = df_est, colour = "black", pch = 19) +
  geom_text(data = df_est, colour = "black",aes(label = label_reformed),hjust =0,vjust=1,size = 3,fontface="italic",family = "Gill Sans") +
  #annotate("text",x = 3,y = 2.75,label = "Policies with\nFavorable MVPF (>0.9)",size = 2,hjust =0,vjust =0) +
  #annotate("text",y = 3.5,x = 2,label = "Policies with\nUnfavorable MVPF (<0.9)",size = 2,hjust = 0, vjust =0)  +
  #ggtitle(paste0("Cost and Benefit Estimates (M=1,000)")) +
  theme(text = element_text(family = "Gill Sans"),
        plot.title = element_text(family = "Gill Sans", face = "plain"),
        plot.background = element_blank(),
        legend.position = "none") +
  geom_point(data = df_psa_medicaid,colour = "blue", alpha = 0.05, pch = 2) + 
  geom_point(data = df_est_medicaid, colour = "blue", pch = 2) +
  geom_text(data = df_est_medicaid, colour = "blue",aes(label = label_reformed),hjust =1,vjust=0,size = 3,fontface="italic",family = "Gill Sans")


p_tmp 
p_tmp %>% ggsave(filename = here("./figures/03_cost-and-benefit-estimates.png"),height = 6, width = 6, units = "in")
```


```{r}
df_psa_params <- 
  params_halton_final %>% 
  map(~(data.frame(.x) %>% 
          rownames_to_column(var = "coef_name") %>% 
          gather(key,value,takeup_coef,cost_coef) %>% 
          unite(key,coef_name,key) %>% 
          spread(key,value))) %>% 
  bind_rows(.id = "iteration")

df_psa_full <- 
  # Subsidy
  df_psa %>% tbl_df() %>% 
    mutate(dQALY_SUBSIDY = mvpf_num, 
           dCOST_SUBSIDY = mvpf_denom) %>% 
    select(iteration,starts_with("dQALY"),starts_with("dCOST"),takeup,C_H) %>% 
  # Medicaid
  left_join(
    df_psa_medicaid %>% tbl_df() %>% 
    mutate(dQALY_MEDICAID = mvpf_num,
           dCOST_MEDICAID = mvpf_denom) %>% 
    select(iteration, starts_with("dQALY"),starts_with("dCOST")),
    "iteration")  %>% 
  # Parameters
    left_join(df_psa_params, "iteration") %>% 
      rename(psa_id = iteration) %>% 
  select(-pop_fpl,-plan_type,-plan_premium,-contains("_coef")) %>% 
  mutate(dQALY_NONE = 0, dCOST_NONE=0)

# df_psa_full  %>% 
#   summarize_at(vars(-psa_id),mean)
parm1<-'gov_incidence'
parm2<-'uninsured_oop_share'
range1<-c(0,1)
range2<-c(0,1)

```


```{r}
p1 <- OneWaySA(indata = df_psa_full %>% select(-contains("SDDUBSIDY")), outcome = "NHB", parm = parm1, range = range1, lambda = 0.8) + 
  #ggtitle("One Way Sensitivity Analysis:\nGovernment Incidence of Uncompensated Care\nMVPF Benchmark = 0.9") + 
  theme(legend.position = "none") +
 xlim(c(0,1.3))  + ylab("E[NWB]") + ylim(c(-1,1)) +
  #theme(plot.title = element_text(family = "Gill Sans",face = "plain")) + 
  theme(plot.title = element_blank()) + 
  theme_tufte_revised() + 
  theme(legend.position = "none", text = element_text(family = "Gill Sans"))+ 
  scale_x_continuous(limits = c(0,1.3), breaks = seq(0,1,0.2))
p1 %>% ggsave(filename = here("./figures/03_one-way-sensitivity_gov-incidence_lambda80.png"))

p2 <- OneWaySA(indata = df_psa_full  %>% select(-contains("SUBSDDDIDY")), outcome = "NHB", parm = parm1, range = range1, lambda = .6) + 
  #ggtitle("One Way Sensitivity Analysis:\nGovernment Incidence of Uncompensated Care\nMVPF Benchmark = 0.2") + 
  theme(legend.position = "none") +
 xlim(c(0,1.3))  + ylab("E[NWB]") + ylim(c(-1,1)) +
  theme(plot.title = element_text(family = "Gill Sans",face = "plain")) + 
  theme_tufte_revised() + 
  theme(legend.position = "none", text = element_text(family = "Gill Sans")) + 
  scale_x_continuous(limits = c(0,1.3), breaks = seq(0,1,0.2))

p2 %>% ggsave(filename = here("./figures/03_one-way-sensitivity_gov-incidence_lambda60.png"))


p3 <-  OneWaySA(indata = df_psa_full  %>% select(-contains("SUBSDDDIDY")), outcome = "NHB", parm = "I", range = c(500,1200), lambda = .6) + 
  #ggtitle("One Way Sensitivity Analysis:\nGovernment Incidence of Uncompensated Care\nMVPF Benchmark = 0.2") + 
  theme(legend.position = "none") +
 xlim(c(500,1200))  + ylab("E[NWB]") + ylim(c(-1,1)) +
  theme(plot.title = element_text(family = "Gill Sans",face = "plain")) + 
  theme_tufte_revised() + 
  theme(legend.position = "none", text = element_text(family = "Gill Sans")) + 
  scale_x_continuous(limits = c(500,1400), breaks = seq(500,1400,200)) + xlab("Insurance Value of Medicaid")

p3 %>% ggsave(filename = here("./figures/03_one-way-sensitivity_insurance-value-medicaid_lambda60.png"))


p1  + ggtitle("One Way Sensitivity Analysis:\nGovernment Incidence of Uncompensated Care\nMVPF Benchmark = 0.9")
p2 + ggtitle("One Way Sensitivity Analysis:\nGovernment Incidence of Uncompensated Care\nMVPF Benchmark = 0.6")
```

```{r}

parm1<-'gov_incidence'
parm2<-'I'
range1<-c(0,1)
range2<-c(100,1250)

#cat("test")
p1 <- 
  TwoWaySA(indata=df_psa_full,outcome="NHB",parm1,parm2,range1,range2,lambda = 0.6) +
  ggthemes::theme_tufte() +
  scale_fill_manual(values = c("black","lightgrey","darkgrey")) +
  #ggtitle("Two-Way Sensitivity Analysis:\nGovernment Incidence of Uncompensated Care\nIncidence on Affluent (eta = 0.5) vs. \nLow-Income (eta = 1)\nMVPF Benchmark = 0.9")  +
  ggthemes::theme_clean() +
  ggtitle(paste0("\u3bB=0.6")) + 
  xlab("Incidence of Uncompensated Care on Government (Share)") + 
  ylab("Insurance Value of Medicaid")+ theme(legend.position = "bottom")

p2 <- 
  TwoWaySA(indata=df_psa_full,outcome="NHB",parm1,parm2,range1,range2,lambda = 0.9) +
  ggthemes::theme_tufte() +
  scale_fill_manual(values = c("black","lightgrey","darkgrey")) +
  #ggtitle("Two-Way Sensitivity Analysis:\nGovernment Incidence of Uncompensated Care\nIncidence on Affluent (eta = 0.5) vs. \nLow-Income (eta = 1)\nMVPF Benchmark = 0.9")  +
  ggthemes::theme_clean() +
  ggtitle(paste0("\u3bB=0.9")) + 
  xlab("Incidence of Uncompensated Care on Government (Share)") + 
  ylab("Insurance Value of Medicaid") + theme(legend.position = "bottom")


p3 <- 
  TwoWaySA(indata=df_psa_full,outcome="NHB",parm1,parm2,range1,range2,lambda = 1.2) +
  ggthemes::theme_tufte() +
  scale_fill_manual(values = c("black","lightgrey","darkgrey")) +
  #ggtitle("Two-Way Sensitivity Analysis:\nGovernment Incidence of Uncompensated Care\nIncidence on Affluent (eta = 0.5) vs. \nLow-Income (eta = 1)\nMVPF Benchmark = 0.9")  +
  ggthemes::theme_clean() +
  ggtitle(paste0("\u3bB=1.2")) + 
  xlab("Incidence of Uncompensated Care on Government (Share)") + 
  ylab("Insurance Value of Medicaid")+ theme(legend.position = "bottom")

p4 <- p1 + p2 + p3

p4 %>% ggsave(filename = here("./figures/03_two-sensitivity_value-medicaid_gov-incidence.png"), width = 20, height =8, units="in")
p4
```

```{r}
lambda_range <- seq(0.0,1.5,0.05)
p_ceac <- 
  CEAC(indata=df_psa_full,lambda_range=lambda_range) + 
  ggthemes::theme_clean() + 
  scale_x_continuous(breaks = seq(0,2,0.1), limits = c(0,2)) + 
  theme(legend.position = "none") + 
  #ggtitle("Policy Acceptability Curve") + 
  ylab("Pr(Cost-Effective)") + 
  geom_vline(aes(xintercept = c(0.2)), lty=2) + 
  annotate("text",x = 0.22, y = 1, label = "Social Welfare Function \n(\U3b4= 3) Benchmark",hjust =0,size =2) +
  geom_vline(aes(xintercept = c(0.88)), lty=2) + 
  annotate("text",x = 0.86, y = 1, label = "EITC\nBenchmark",hjust =1,size = 2) +
  scale_colour_brewer(palette = "Set1")+
  scale_colour_grey( start = 0, end = .5, na.value = "red") + 
   theme(plot.title = element_text(family = "Gill Sans",face = "plain")) + 
  theme_tufte_revised() + 
  theme(legend.position = "none", text = element_text(family = "Gill Sans")) +
    xlab(expression("Policy Adoption Threshold [Benefit/Cost]" (lambda))) 
  
  #scale_x_continuous(limits = c(0,1.3), breaks = seq(0,1,0.2))
p_ceac + ggtitle("Policy Acceptability Frontier")
p_ceac %>% ggsave(filename=here("./figures/03_policy-acceptability-frontier.png"), width = 7, height = 4, units = "in")
```


```{r}
# TornadoDiag(indata = df_psa_full %>% select(-contains("NOSUBSIDY")), outcome = "NHB",lambda = 0.88) + 
#   xlab("Expected NWB") + 
#   geom_vline(aes(xintercept = 0), colour = "red")
```

```{r}
indata <- df_psa_full 

evppi.res <- list()

for (lambda in seq(0.2,1.5,0.1)) {
  # Get Outcome
  #cat(lambda)
  
  lhs <- indata %>% select(psa_id,contains("dQALY"),contains("dCOST")) %>%
    mutate(psa_id=row_number()) %>% 
    reshape2::melt(id.vars='psa_id') %>%
    tidyr::separate(variable,c("outcome","strategy"),"_") %>%
    reshape2::dcast(psa_id+strategy~outcome) %>%
    mutate(NHB = dQALY- dCOST * lambda ,
           NMB = dQALY - dCOST *lambda)
  nmb <- lhs %>% select(psa_id,strategy,NMB) %>% 
    mutate(variable="NMB") %>% reshape2::dcast(psa_id~variable+strategy,value.var="NMB") %>% 
    select(-psa_id)
  
  # Get Parameters
  theta_tmp <- indata %>% select(-contains("dQALY"),-contains("dCOST"),
                             -contains("NMB"),-contains("NHB"),-psa_id)
  
  # Get all the varying parameters
  theta_var <- theta_tmp %>% summarise_all(sd) %>% 
    gather(key,value) %>% 
    filter(value>0) %>% 
    pull(key)
  
  theta <- theta_tmp %>% select(theta_var)
  
  ## Number of simulations
  n.sim        <- nrow(nmb)
  ## Number of strategies
  n.strategies <- ncol(nmb)
  
  ### Load required packages and functions
  ## For column and row stats
  # library(matrixStats)  
  # ## To fit spline models
  # library(mgcv) # version 1.8-17
  ## Functions to calculate the conditional loss by computing the 
  ## preposterior of each of the basis functions of the GAM model
  #source("../reference/metamodeling/supplemental-material/GA_functions.R")
  
  ### Find optimal strategy (d*) based on the highest expected NMB
  d.star <- which.max(colMeans(nmb))
  #d.star
  
  ### Define the Loss matrix
  loss <- nmb - nmb[, d.star]
  
  ### EVPI
  evpi <- mean(rowMaxs(as.matrix(loss)))
  #evpi
  
  #========================#
  #### Single parameter ####
  #========================#
  ### Generate linear metamodel of one parameter for each opportunity loss
  ## Selected parameter for EVPPI & EVSI
  
  evppi.res[[paste0(lambda)]] <- data.frame(parameter = names(theta)) %>% mutate(evppi = NA)
  
  for (sel.param in seq(ncol(theta))) {
    
    #cat("\n")
    #cat(paste0("  ",sel.param))
    
    lmm1 <- gam(as.formula(paste("loss[, 1] ~ s(", colnames(theta)[sel.param], ")")),
                data = theta)
    lmm2 <- gam(as.formula(paste("loss[, 2] ~ s(", colnames(theta)[sel.param], ")")),
                data = theta)
    lmm3 <- gam(as.formula(paste("loss[, 3] ~ s(", colnames(theta)[sel.param], ")")),
                data = theta)
    #lmm4 <- gam(as.formula(paste("loss[, 4] ~ s(", colnames(theta)[sel.param], ")")),
                #data = theta)  
    
    #### Compute EVPPI on one parameter ####
    ## Compute estimated losses
    loss.hat <- cbind(lmm1$fitted, lmm2$fitted, lmm3$fitted)#,lmm4$fitted)
    
    ### Apply EVPPI equation
    evppi.res[[paste0(lambda)]][sel.param,]$evppi <- mean(rowMaxs(loss.hat))
    #cat("\n")
  }
} 
df_evppi <- evppi.res %>% bind_rows(.id = "lambda") %>% tbl_df() %>% 
  mutate(paramter = paste0(parameter)) %>% 
  group_by(parameter) %>% 
  mutate(max_evppi = max(evppi)) %>% 
  filter(max_evppi > 0.005) %>% 
  ungroup()

df_evppi_label <- 
  df_evppi %>% 
  select(parameter) %>% 
  unique() %>% 
  mutate(parameter = paste0(parameter)) %>% 
  mutate(tmp = params_lut[parameter]) %>% 
  mutate(label = paste0(parameter,": ",tmp))
  
library(RColorBrewer)
myColors <- rep("black",nrow(df_evppi_label)) #brewer.pal(12,"Set3")
names(myColors) <- df_evppi_label$parameter
colScale <- scale_colour_manual(name = "",values = myColors, labels = df_evppi_label$label)
#lineScale <- scale_linetype_manual(values = seq(1,nrow(df_evppi_label), labels = df_evppi_label$label))

p_evppi <- 
  df_evppi %>% 
  #filter(lambda>0.8 & lambda<1.2) %>% 
  ggplot(aes(x = lambda, y = evppi, group = parameter, colour = parameter)) + 
  geom_line() + 
  # geom_text(data = 
  # evppi.res %>% bind_rows(.id = "lambda") %>% filter(evppi>0), aes(label = round(evppi,3)),size =3) +
  
  theme(legend.position = "none") +
  theme_tufte_revised() + #theme(legend.position = "none") + 
  ylab("Expected Value of Partial Perfect Information") + 
     xlab(expression("Policy Adoption Threshold [Benefit/Cost]" (lambda)))  + 
  theme(text = element_text(family = "Gill Sans")) +
  geom_dl(aes(label = parameter), method = list("top.bumpup",cex =0.9,fontfamily = "Gill Sans"))  + 
  colScale + 
  theme(legend.position = "bottom")
  #ggtitle("Expected Value of Perfect Information by Parameter")
p_evppi + ggtitle("Degree of Parameter Leverage in Affecting Policy Adoption,\nby MVPF Threshold Value")
p_evppi %>% ggsave(filename = here("./figures/03_evppi.png"),height = 6, width = 12, units = "in")
```


```{r}
p1 <- p_ceac + ggtitle("Policy Acceptability Frontier")
p2 <- p_evppi + ggtitle("Degree of Parameter Leverage in Affecting Policy Adoption,\nby MVPF Threshold Value")

p3 <- p1 + p2 + plot_layout(nrow=1)
p3 %>% ggsave(filename = here("./figures/99_summary-of-results.png"),height = 6, width = 12, units = "in")

```
